ref: Move changed tests to version tests
