+++ commands/changed/__tests__/changed-command.test.js
@@ -4,7 +4,6 @@
 jest.unmock("@lerna/collect-updates");
 
 const path = require("path");
-const fs = require("fs");
 const touch = require("touch");
 
 // mocked modules
@@ -15,8 +14,6 @@ const initFixture = require("@lerna-test/init-fixture")(__dirname);
 const gitAdd = require("@lerna-test/git-add");
 const gitCommit = require("@lerna-test/git-commit");
 const gitTag = require("@lerna-test/git-tag");
-const gitCheckout = require("@lerna-test/git-checkout");
-const gitMerge = require("@lerna-test/git-merge");
 const updateLernaConfig = require("@lerna-test/update-lerna-config");
 
 // file under test
@@ -42,26 +39,6 @@ const setupGitChanges = async (cwd, filePaths) => {
   await Promise.all(filePaths.map(fp => touch(path.join(cwd, fp))));
   await gitAdd(cwd, "-A");
   await gitCommit(cwd, "Commit");
-};
-
-const setupGitChangesWithBranch = async (cwd, masterPaths, branchPaths) => {
-  await gitTag(cwd, "v1.0.0");
-  await Promise.all(masterPaths.map(fp => fs.appendFileSync(path.join(cwd, fp), "1")));
-  await gitAdd(cwd, "-A");
-  await gitCommit(cwd, "Commit");
-  // Create release branch
-  await gitCheckout(cwd, ["-b", "release/v1.0.1"]);
-  // Change in release branch
-  await Promise.all(branchPaths.map(fp => fs.appendFileSync(path.join(cwd, fp), "1")));
-  await gitAdd(cwd, "-A");
-  await gitCommit(cwd, "Bump");
-  await gitTag(cwd, "v1.0.1");
-  await gitCheckout(cwd, ["master"]);
-  await gitMerge(cwd, ["--no-ff", "release/v1.0.1"]);
-  // Commit after merge
-  await Promise.all(masterPaths.map(fp => fs.appendFileSync(path.join(cwd, fp), "1")));
-  await gitAdd(cwd, "-A");
-  await gitCommit(cwd, "Commit2");
 };
 
 describe("ChangedCommand", () => {
@@ -184,44 +161,6 @@ __TEST_ROOTDIR__/packages/package-2:package-2:1.0.0
 __TEST_ROOTDIR__/packages/package-3:package-3:1.0.0
 __TEST_ROOTDIR__/packages/package-4:package-4:1.0.0
 __TEST_ROOTDIR__/packages/package-5:package-5:1.0.0:PRIVATE
-`);
-    });
-
-    it("should list changes to package-4", async () => {
-      const testDir = await initFixture("basic");
-
-      await setupGitChangesWithBranch(
-        testDir,
-        ["packages/package-2/random-file"],
-        ["packages/package-4/random-file"]
-      );
-      // Without --include-merged-tags we receive all changes since the last tag on master
-      // in this case it's v1.0.0, this includes changes to package-4 which was released
-      // in the release branch with v1.0.1
-      await lernaChanged(testDir)();
-
-      expect(output.logged()).toMatchInlineSnapshot(`
-package-2
-package-3
-package-4
-`);
-    });
-
-    it("should not list changes to package-4 with --include-merged-tags", async () => {
-      const testDir = await initFixture("basic");
-
-      await setupGitChangesWithBranch(
-        testDir,
-        ["packages/package-2/random-file"],
-        ["packages/package-4/random-file"]
-      );
-      // With --include-merged-tags we correctly detect that v1.0.1 was already tagged
-      // and merged. We no longer want to receive package-4.
-      await lernaChanged(testDir)("--include-merged-tags");
-
-      expect(output.logged()).toMatchInlineSnapshot(`
-package-2
-package-3
 `);
     });
   });
