+++ commands/changed/command.js
@@ -15,13 +15,7 @@ exports.describe = "List local packages that have changed since the last tagged 
 exports.builder = yargs => {
   listable.options(yargs);
 
-  const newYargs = versionOptions(yargs, "changed");
-  newYargs.option("include-merged-tags", {
-    describe: "Also include tags from merged branches",
-    type: "boolean",
-    defaultDescription: "false",
-  });
-  return newYargs;
+  return versionOptions(yargs, "changed");
 };
 
 exports.handler = function handler(argv) {
