+++ commands/version/__tests__/__fixtures__/basic/lerna.json
@@ -1,0 +1,3 @@
+{
+  "version": "1.0.0"
+}
