+++ commands/version/__tests__/__fixtures__/basic/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "independent"
+}
