+++ commands/version/__tests__/__fixtures__/basic/packages/package-5/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "package-5",
+  "version": "1.0.0",
+  "private": true,
+  "dependencies": {
+    "package-1": "^1.0.0"
+  }
+}
