+++ utils/describe-ref/__tests__/describe-ref.test.js
@@ -49,7 +49,7 @@ describe("describeRef()", () => {
     const newArgs = [...DEFAULT_ARGS];
     newArgs.pop();
     newArgs.push("--tags");
-    expect(childProcess.exec).lastCalledWith("git", newArgs, {});
+    expect(childProcess.exec).toHaveBeenLastCalledWith("git", newArgs, {});
   });
 });
